requirements.txt
--extra-index-url https://test.pypi.org/simple/
python-telegram-bot~=12.0.0
analner~=0.0.11
python-dotenv~=0.10.3
redis~=3.0.1
pyowm~=2.10.0
git+https://github.com/40min/bash-org-sentiment-analyser.git#egg=bash-org-sentiment-analyser
dialogflow~=0.6.0
